problems/24.Swap-Nodes-in-Pairs/24.Swap Nodes in Pairs.go
package leetcode

import (
	"leetcode/structs"
)

/**
 * Definition for singly-linked list.
 * type ListNode struct {
 *     Val int
 *     Next *ListNode
 * }
 */
func swapPairs(head *structs.ListNode) *structs.ListNode {

	dummy := &structs.ListNode{Next: head}
	for pt := dummy; pt != nil && pt.Next != nil && pt.Next.Next != nil; {
		// dummy -> t1 -> t2 -> t3  = t1 -> t2 -> t3 -> t1
		pt, pt.Next, pt.Next.Next, pt.Next.Next.Next = pt.Next, pt.Next.Next, pt.Next.Next.Next, pt.Next
	}
	return dummy.Next
}
